# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fronei API",
    "version": "0.1.0",
    "description": "REST API for Fronei \u2014 AI-driven enterprise transformation assistant. Covers chat/conversations, research runs, document extraction, memory & writing-voice (twin profile), model routing/policy, analytics, and admin operations (user management, roles, providers, routing policy, audit logs, system status).",
    "contact": {
      "name": "Fronei API"
    }
  },
  "paths": {
...
    "/memory": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "List Memories",
        "operationId": "list_memories_memory_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoryItem"
                  },
                  "title": "Response List Memories Memory Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Create Memory",
        "operationId": "create_memory_memory_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Clear Memories",
        "operationId": "clear_memories_memory_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Confirm"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/{memory_id}": {
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete Memory",
        "operationId": "delete_memory_memory__memory_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/twin-profile": {
      "get": {
        "tags": [
          "twin-profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_twin_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwinProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/twin-profile/prefs": {
      "put": {
        "tags": [
          "twin-profile"
        ],
        "summary": "Update Prefs",
        "operationId": "update_prefs_twin_profile_prefs_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwinProfilePrefsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwinProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/twin-profile/samples": {
      "get": {
        "tags": [
          "twin-profile"
        ],
        "summary": "List Samples",
        "operationId": "list_samples_twin_profile_samples_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WritingSampleOut"
                  },
                  "type": "array",
                  "title": "Response List Samples Twin Profile Samples Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "twin-profile"
        ],
        "summary": "Add Sample",
        "operationId": "add_sample_twin_profile_samples_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritingSampleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingSampleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/twin-profile/samples/{sample_id}": {
      "delete": {
        "tags": [
          "twin-profile"
        ],
        "summary": "Delete Sample",
        "operationId": "delete_sample_twin_profile_samples__sample_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/twin-profile/extract": {
      "post": {
        "tags": [
          "twin-profile"
        ],
        "summary": "Trigger Extraction",
        "description": "Manually trigger fingerprint re-extraction from all samples.",
        "operationId": "trigger_extraction_twin_profile_extract_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Trigger Extraction Twin Profile Extract Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FingerprintOut": {
        "properties": {
          "sentence_length": {
            "type": "string",
            "title": "Sentence Length",
            "default": "medium"
          },
          "formality": {
            "type": "string",
            "title": "Formality",
            "default": "professional"
          },
          "directness": {
            "type": "string",
            "title": "Directness",
            "default": "high"
          },
          "hedging": {
            "type": "string",
            "title": "Hedging",
            "default": "low"
          },
          "structure": {
            "type": "string",
            "title": "Structure",
            "default": "mixed"
          },
          "technical_depth": {
            "type": "string",
            "title": "Technical Depth",
            "default": "high"
          },
          "preferred_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Phrases",
            "default": []
          },
          "forbidden_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Forbidden Phrases",
            "default": []
          },
          "avoid_patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoid Patterns",
            "default": []
          },
          "signature_patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signature Patterns",
            "default": []
          },
          "tone_by_audience": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tone By Audience",
            "default": {}
          }
        },
        "type": "object",
        "title": "FingerprintOut",
        "description": "Structured fingerprint extracted from writing samples."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MemoryCreate": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Content"
          },
          "category": {
            "type": "string",
            "maxLength": 64,
            "title": "Category",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MemoryCreate"
      },
      "MemoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "source_conversation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Conversation Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "category",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryItem"
      },
...
      "TwinProfileOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "fingerprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FingerprintOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "rewrite_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewrite Prompt"
          },
          "prefs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Prefs",
            "default": {}
          },
          "extracted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted At"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "TwinProfileOut"
      },
      "TwinProfilePrefsUpdate": {
        "properties": {
          "preferred_phrases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Phrases"
          },
          "forbidden_phrases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forbidden Phrases"
          },
          "tone_by_audience": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone By Audience"
          }
        },
        "type": "object",
        "title": "TwinProfilePrefsUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WritingSampleIn": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 50,
            "title": "Content"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "WritingSampleIn"
      },
      "WritingSampleOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "label",
          "char_count",
          "created_at"
        ],
        "title": "WritingSampleOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ]
}